🧰: combine code search into one module to fix circular imports
